feat: Update blue color shade in Colors.ts
This commit updates the blue color shade in the Colors.ts file from '#002073' to '#1C8EF4'. The new shade provides a more vibrant and appealing look to the application's profile screen.

chore: Update AuthInputField style in AuthInput.tsx

This commit updates the style of the AuthInputField component in the AuthInput.tsx file. The commented out style code is replaced with a new style that includes a border width of 1, a width of 100%, and a height of 45. This change enhances the visual appearance of the input field.

feat: Add ActiveOrderdetails screen

This commit adds the ActiveOrderdetails screen to the Admin section of the application. The screen displays the details of an active order and provides relevant information to the admin user.

feat: Add Slider component

This commit adds the Slider component to the ActiveOrder folder in the Admin section of the application. The Slider component is responsible for displaying a scrollable list of banner images. It accepts an array of banner images as a prop and allows the user to swipe through them.

feat: Optimize images

This commit optimizes the images in the project using ImgBot. The total size of the images is reduced from 2,348.84kb to 1,992.29kb, resulting in a 15.18% reduction in size. Individual images are also optimized, with reductions ranging from 4.43% to 23.8%.

feat: Add expo-clipboard and @sentry/react-native dependencies

This commit adds the expo-clipboard and @sentry/react-native dependencies to the project's package.json file. These dependencies are necessary for clipboard functionality and error tracking using Sentry in the application.

chore: Update app.json, Images.tsx, SubmitBtn, and AuthInputField

This commit updates the app.json file, the Images.tsx file, and the SubmitBtn and AuthInputField components. The changes include adding a busy prop to the SubmitBtn component and a maxLength prop to the AuthInputField component. Additionally, the app.json file is modified to reflect the updated configuration.